+++ CHANGELOG.md
@@ -15,7 +15,8 @@
 - Add `App.jsx` component insertion point @sneridagh
 - Lazy load `EventView.jsx`, splitting out all its big dependencies (Luxon) along with it @sneridagh
 - Lazy load `pretty` library, used in the HTML block @sneridagh
-- Lazy load Raven library @sneridagh
+- Lazy load `Raven` library @sneridagh
+- Lazy load `React Select` library @sneridagh
 
 ### Bugfix